# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plus Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppTag = (typeof AppTag)[keyof typeof AppTag];

export const AppTag = {
  tweaked: "tweaked",
  modded: "modded",
  hacked: "hacked",
  new: "new",
  hot: "hot",
} as const;

export interface App {
  id: number;
  name: string;
  description?: string;
  icon: string;
  categoryId: number;
  categoryName: string;
  tag: AppTag;
  version?: string;
  size?: string;
  downloads?: number;
  isFeatured?: boolean;
  isHot?: boolean;
  createdAt?: string;
}

export interface AppListResponse {
  apps: App[];
  total: number;
  page?: number;
  limit?: number;
}

export interface Category {
  id: number;
  name: string;
  nameAr?: string;
  icon: string;
  appCount?: number;
}

export interface CategoryListResponse {
  categories: Category[];
}

export interface Plan {
  id: number;
  name: string;
  nameAr?: string;
  price: number;
  currency: string;
  duration: string;
  features: string[];
  excludedFeatures?: string[];
  isPopular?: boolean;
}

export interface PlanListResponse {
  plans: Plan[];
}

export interface ActivateSubscriptionRequest {
  code: string;
  udid?: string;
}

export interface ActivationResponse {
  success: boolean;
  message: string;
  expiresAt?: string;
}

export type CreateAppRequestTag =
  (typeof CreateAppRequestTag)[keyof typeof CreateAppRequestTag];

export const CreateAppRequestTag = {
  tweaked: "tweaked",
  modded: "modded",
  hacked: "hacked",
  new: "new",
  hot: "hot",
} as const;

export interface CreateAppRequest {
  name: string;
  description?: string;
  icon: string;
  categoryId: number;
  tag: CreateAppRequestTag;
  version?: string;
  size?: string;
  isFeatured?: boolean;
  isHot?: boolean;
}

export interface CreateCategoryRequest {
  name: string;
  nameAr?: string;
  icon: string;
}

export interface CreatePlanRequest {
  name: string;
  nameAr?: string;
  price: number;
  currency: string;
  duration: string;
  features: string[];
  excludedFeatures?: string[];
  isPopular?: boolean;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface AdminStats {
  totalApps: number;
  totalCategories: number;
  totalSubscriptions: number;
  activeSubscriptions: number;
}

export type ListAppsParams = {
  categoryId?: number;
  filter?: ListAppsFilter;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListAppsFilter =
  (typeof ListAppsFilter)[keyof typeof ListAppsFilter];

export const ListAppsFilter = {
  all: "all",
  hot: "hot",
  new: "new",
  tweaked: "tweaked",
  modded: "modded",
  hacked: "hacked",
} as const;

export type AdminListAppsParams = {
  page?: number;
  limit?: number;
};