# lib/api-zod/src/generated/types/activateSubscriptionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plus Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivateSubscriptionRequest {
  code: string;
  udid?: string;
}