# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plus Store API
 * OpenAPI spec version: 0.1.0
 */
import type { AppTag } from "./appTag";

export interface App {
  id: number;
  name: string;
  description?: string;
  icon: string;
  categoryId: number;
  categoryName: string;
  tag: AppTag;
  version?: string;
  size?: string;
  downloads?: number;
  isFeatured?: boolean;
  isHot?: boolean;
  createdAt?: Date;
}