# lib/api-zod/src/generated/types/categoryListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plus Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface CategoryListResponse {
  categories: Category[];
}