# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plus Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  nameAr?: string;
  price: number;
  currency: string;
  duration: string;
  features: string[];
  excludedFeatures?: string[];
  isPopular?: boolean;
}